refactor(template): 添加模板卡片、缩略图和用户模板管理功能
- 新增模板卡片组件，支持预览、标签和操作按钮。
- 新增模板缩略图组件，展示模板的预览效果。
- 实现用户模板管理功能，包括删除、编辑和发布模板。
- 添加模板相关常量和上下文管理，优化模板数据处理。
- 更新模板编辑器和工作台的状态管理逻辑。

Closes #123
