+++ src/components/resume/runtime/TemplateRuntimeProviders.tsx
@@ -1,10 +1,10 @@
 import type { PropsWithChildren } from 'react'
 import type { ResolvedTemplateManifest } from '@/lib/resume-template/schema'
 import type { ResumeAppearanceConfig } from '@/lib/schema'
-import type { TemplateResumeData } from '@/pages/template/components/resume-data-context'
+import type { TemplateResumeData } from '@/pages/template/context/resume-data-context'
 import { useResumeStyles } from '@/hooks/use-resume-styles'
-import { ResumeContextProvider } from '@/pages/template/components/resume-context'
-import { TemplateResumeDataProvider } from '@/pages/template/components/resume-data-context'
+import { ResumeContextProvider } from '@/pages/template/context/resume-context'
+import { TemplateResumeDataProvider } from '@/pages/template/context/resume-data-context'
 
 export function TemplateRuntimeProviders({
   children,
