+++ src/components/resume/runtime/renderers/ApplicationInfoRenderer.tsx
@@ -1,4 +1,4 @@
-import { useTemplateResumeData } from '@/pages/template/components/resume-data-context'
+import { useTemplateResumeData } from '@/pages/template/context/resume-data-context'
 import { RuntimeSection } from './shared'
 import { useRuntimeStyles } from './utils'
