+++ src/pages/template/components/basic/index.tsx
@@ -1,7 +1,7 @@
 import type { ResumeAppearanceConfig } from '@/lib/schema'
 import { ResumeTemplateRuntime } from '@/components/resume/runtime/ResumeTemplateRuntime'
 import { getBuiltInTemplateManifest } from '@/lib/resume-template/runtime'
-import { useTemplateResumeData } from '../resume-data-context'
+import { useTemplateResumeData } from '@/pages/template/context/resume-data-context'
 
 interface BasicResumeProps {
   appearance?: Partial<ResumeAppearanceConfig> | null
