+++ src/pages/template/components/editor/preview-resume-select.tsx
@@ -1,11 +1,5 @@
 import type { TemplatePreviewResumeOption } from '../../hooks/use-template-preview-resume'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 
 interface TemplatePreviewResumeSelectProps {
   options: TemplatePreviewResumeOption[]
