+++ src/pages/template/store/editor.ts
