+++ src/pages/template/store/index.ts
@@ -1,0 +1,3 @@
+export { default as useTemplateEditorStore } from './editor'
+export { default as useTemplateWorkbenchStore } from './workbench'
+export type { TemplateWorkbenchMode, TemplateWorkbenchSource, TemplateWorkbenchTab } from './workbench'
