+++ src/pages/template/types.ts
@@ -1,0 +1,1 @@
+export type { TemplateWorkbenchMode, TemplateWorkbenchSource, TemplateWorkbenchTab } from './store/workbench'
